# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.1
paths:
...
  /api/v1/equipment-mods:
    get:
      tags:
      - equipment-mods
      summary: List Equip Mods
      operationId: list_equip_mods_api_v1_equipment_mods_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filter by category: weapon, engine, shield, ship'
          title: Category
        description: 'Filter by category: weapon, engine, shield, ship'
      - name: stat
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by stat type e.g. damage, cooling
          title: Stat
        description: Filter by stat type e.g. damage, cooling
      - name: quality
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: "Filter by quality tier 1\u20133"
          title: Quality
        description: "Filter by quality tier 1\u20133"
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EquipModSummary'
                title: Response List Equip Mods Api V1 Equipment Mods Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/equipment-mods/{ware_id}:
    get:
      tags:
      - equipment-mods
      summary: Get Equip Mod
      operationId: get_equip_mod_api_v1_equipment_mods__ware_id__get
      parameters:
      - name: ware_id
        in: path
        required: true
        schema:
          type: string
          title: Ware Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipModDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/diplomacy/actions:
    get:
      tags:
      - diplomacy
      summary: List Diplo Actions
      description: List all diplomatic agent actions.
      operationId: list_diplo_actions_api_v1_diplomacy_actions_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filter by category: negotiation | espionage | interference'
          title: Category
        description: 'Filter by category: negotiation | espionage | interference'
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Hidden
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiploAction'
                title: Response List Diplo Actions Api V1 Diplomacy Actions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/diplomacy/actions/{action_id}:
    get:
      tags:
      - diplomacy
      summary: Get Diplo Action
      operationId: get_diplo_action_api_v1_diplomacy_actions__action_id__get
      parameters:
      - name: action_id
        in: path
        required: true
        schema:
          type: string
          title: Action Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiploAction'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/diplomacy/gifts:
    get:
      tags:
      - diplomacy
      summary: List Diplo Gifts
      description: List faction gift preferences. Filter by faction_id to find what
        a faction likes.
      operationId: list_diplo_gifts_api_v1_diplomacy_gifts_get
      parameters:
      - name: faction_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiploGift'
                title: Response List Diplo Gifts Api V1 Diplomacy Gifts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/drops/lists:
    get:
      tags:
      - drops
      summary: List Drop Lists
      description: List all named drop tables.
      operationId: list_drop_lists_api_v1_drops_lists_get
      parameters:
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ship | lockbox | asteroid | crystal | story | masstraffic |
            other
          title: Category
        description: ship | lockbox | asteroid | crystal | story | masstraffic | other
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DropList'
                title: Response List Drop Lists Api V1 Drops Lists Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/drops/lists/{list_id}:
    get:
      tags:
      - drops
      summary: Get Drop List
      description: Get a specific drop table with all its resolved ware entries.
      operationId: get_drop_list_api_v1_drops_lists__list_id__get
      parameters:
      - name: list_id
        in: path
        required: true
        schema:
          type: string
          title: List Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DropListDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/drops/wares/{ware_id}:
    get:
      tags:
      - drops
      summary: Get Ware Drop Sources
      description: Find all drop tables that can yield a specific ware (reverse lookup).
      operationId: get_ware_drop_sources_api_v1_drops_wares__ware_id__get
      parameters:
      - name: ware_id
        in: path
        required: true
        schema:
          type: string
          title: Ware Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DropSource'
                title: Response Get Ware Drop Sources Api V1 Drops Wares  Ware Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    BribeWare:
      properties:
        ware_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware Id
        ware_tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Ware Tags
        amount:
          anyOf:
          - type: integer
          - type: 'null'
          title: Amount
      additionalProperties: false
      type: object
      required:
      - ware_id
      - ware_tags
      - amount
      title: BribeWare
...
    DiploAction:
      properties:
        action_id:
          type: string
          title: Action Id
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        hidden:
          type: integer
          title: Hidden
        cost_influence:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cost Influence
        cost_money:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cost Money
        success_chance:
          anyOf:
          - type: integer
          - type: 'null'
          title: Success Chance
        duration_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Sec
        cooldown_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cooldown Sec
        agent_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Type
        agent_experience:
          anyOf:
          - type: integer
          - type: 'null'
          title: Agent Experience
        risk:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk
        bribe_wares:
          items:
            $ref: '#/components/schemas/BribeWare'
          type: array
          title: Bribe Wares
      additionalProperties: false
      type: object
      required:
      - action_id
      - category
      - name
      - description
      - hidden
      - cost_influence
      - cost_money
      - success_chance
      - duration_sec
      - cooldown_sec
      - agent_type
      - agent_experience
      - risk
      - bribe_wares
      title: DiploAction
    DiploGift:
      properties:
        ware_id:
          type: string
          title: Ware Id
        faction_id:
          type: string
          title: Faction Id
      additionalProperties: false
      type: object
      required:
      - ware_id
      - faction_id
      title: DiploGift
    DropEntry:
      properties:
        ware_id:
          type: string
          title: Ware Id
        spawn_chance:
          anyOf:
          - type: integer
          - type: 'null'
          title: Spawn Chance
        item_chance:
          anyOf:
          - type: number
          - type: 'null'
          title: Item Chance
        min_amount:
          type: integer
          title: Min Amount
        max_amount:
          type: integer
          title: Max Amount
        source_basket:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Basket
      additionalProperties: false
      type: object
      required:
      - ware_id
      - spawn_chance
      - item_chance
      - min_amount
      - max_amount
      - source_basket
      title: DropEntry
    DropList:
      properties:
        list_id:
          type: string
          title: List Id
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      additionalProperties: false
      type: object
      required:
      - list_id
      - category
      title: DropList
    DropListDetail:
      properties:
        list_id:
          type: string
          title: List Id
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        wares:
          items:
            $ref: '#/components/schemas/DropEntry'
          type: array
          title: Wares
      additionalProperties: false
      type: object
      required:
      - list_id
      - category
      - wares
      title: DropListDetail
    DropSource:
      properties:
        list_id:
          type: string
          title: List Id
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        spawn_chance:
          anyOf:
          - type: integer
          - type: 'null'
          title: Spawn Chance
        item_chance:
          anyOf:
          - type: number
          - type: 'null'
          title: Item Chance
        min_amount:
          type: integer
          title: Min Amount
        max_amount:
          type: integer
          title: Max Amount
        source_basket:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Basket
      additionalProperties: false
      type: object
      required:
      - list_id
      - category
      - spawn_chance
      - item_chance
      - min_amount
      - max_amount
      - source_basket
      title: DropSource
    EquipModDetail:
      properties:
        ware_id:
          type: string
          title: Ware Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        shortname:
          anyOf:
          - type: string
          - type: 'null'
          title: Shortname
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        stat:
          anyOf:
          - type: string
          - type: 'null'
          title: Stat
        quality:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quality
        min_factor:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Factor
        max_factor:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Factor
        price_avg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Avg
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        price_min:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Min
        price_max:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Max
        production_time:
          anyOf:
          - type: number
          - type: 'null'
          title: Production Time
      additionalProperties: false
      type: object
      required:
      - ware_id
      - name
      - shortname
      - category
      - stat
      - quality
      - min_factor
      - max_factor
      - price_avg
      - description
      - price_min
      - price_max
      - production_time
      title: EquipModDetail
    EquipModSummary:
      properties:
        ware_id:
          type: string
          title: Ware Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        shortname:
          anyOf:
          - type: string
          - type: 'null'
          title: Shortname
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        stat:
          anyOf:
          - type: string
          - type: 'null'
          title: Stat
        quality:
          anyOf:
          - type: integer
          - type: 'null'
          title: Quality
        min_factor:
          anyOf:
          - type: number
          - type: 'null'
          title: Min Factor
        max_factor:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Factor
        price_avg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Price Avg
      additionalProperties: false
      type: object
      required:
      - ware_id
      - name
      - shortname
      - category
      - stat
      - quality
      - min_factor
      - max_factor
      - price_avg
      title: EquipModSummary
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...